fixing a few mismatching and conflicting bit defs
HAS_VERT_UNCERTAINITY needs to be 0x20 as it is also defined
so in HIDL. But 0x20 conflicts with LOCATION_HAS_SOURCE_INFO,
so moving it into LOC_GPS_LOCATION_HAS_* bit set.
At some point we need to translate the bits instead, but that
would call for a larger change.

CRs-Fixed: 2817983
Change-Id: I9e631dc87127cde6f51e7bea74e472afe0f8607f