Restoring AssistedGpsDataItem support
Receiving ASSISTED_GPS enabled/disabled events from OSAgent
and using it to decide the SUPL mode to be used for tracking
sessions.
GnssConfiguration module from the framework also sends the
SUPL_MODE down via IGnssConfiguration setSuplMode, which takes
the CarrierConfig as well and gps.conf into account.
Finally in GnssAdapter, we consider both the ASSISTED_GPS
dataitem and the SUPL_MODE from GnssConfig to decide the
SUPL MODE to be set for the tracking session.
This takes effect for any client who sets the SUPL mode as
UNKNOWN.

CRs-Fixed: 4149309
Change-Id: I89498a47c9c37fb043fb2b1c6ab770ff50320d82